Implemented pipeline actor factory test suite
